Copied over David's changes from TS rewrite
